azure-table-cleaner/table_cleaner/requirements.txt
azure-functions
azure-data-tables>=12.0.0
